refactor decond main.rb to have nex main.rb and option.rb to show_option method
